requirements.txt
PyQt5~=5.15.7
Pillow~=9.2.0
XlsxWriter~=3.0.3